virtio: Enable the PUBLISH_USED feature by default for all devices
RH-Author: Amit Shah <amit.shah@redhat.com>
Message-id: <c71c78a5dcd5d5782a4c95875dafd21dc84f9504.1276833209.git.amit.shah@redhat.com>
Patchwork-id: 9969
O-Subject: [RHEL6 PATCH V2] virtio: Enable the PUBLISH_USED feature by default
	for all devices
Bugzilla: 602417
RH-Acked-by: Michael S. Tsirkin <mst@redhat.com>
RH-Acked-by: Juan Quintela <quintela@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Alex Williamson <alex.williamson@redhat.com>

When both the host and the guest support the PUBLISH_USED feature, the
guest lets the host know how far it is in its consumption of the buffers
in virtqueues. Based on this knowledge, the guest may not raise an
interrupt on new virtqueue pushes to save the guest the interrupt work,
as the guest anyway would be processing the data it has already queued
up.

This reduces guest interrupts and vmexits.

This was only enabled for virtio-net earlier, as virtio-blk had
exhibited some performance loss. Later tests reveal no performance
degradation for virtio-blk as well as virtio-serial. So enable for all
the devices now.

And now that the PUBLISH_USED feature is available for all virtio
devices, remove the net-specific define.

Signed-off-by: Amit Shah <amit.shah@redhat.com>
---
 hw/virtio-net.h |    2 --
 hw/virtio.h     |    4 +++-
 2 files changed, 3 insertions(+), 3 deletions(-)

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>