QError: Introduce QERR_DEVICE_INIT_FAILED_2
RH-Author: Luiz Capitulino <lcapitulino@redhat.com>
Message-id: <1277492961-24090-2-git-send-email-lcapitulino@redhat.com>
Patchwork-id: 10259
O-Subject: [PATCH 1/2] QError: Introduce QERR_DEVICE_INIT_FAILED_2
Bugzilla: 596279
RH-Acked-by: Alex Williamson <alex.williamson@redhat.com>
RH-Acked-by: Kevin Wolf <kwolf@redhat.com>
RH-Acked-by: Juan Quintela <quintela@redhat.com>

This is an hacky extension to existing DeviceInitFailed error which
provides more end-user information regarding the error cause.

The main motivation is to allow the device assigment driver to report
more specific errors when it's unable to assign an IRQ for the device
being attached.

This is really an exception, which is justified by the fact that
libvirt is unable to detect this problem and this can turn out to be
a source of bug reports if we use a generic error message for it,
as explained in this BZ comment:

 https://bugzilla.redhat.com/show_bug.cgi?id=596279#c6

Signed-off-by: Luiz Capitulino <lcapitulino@redhat.com>
---
 qerror.c |    4 ++++
 qerror.h |    4 ++++
 2 files changed, 8 insertions(+), 0 deletions(-)

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>