ramblocks: Make use of DeviceState pointer and BusInfo.get_dev_path
RH-Author: Alex Williamson <alex.williamson@redhat.com>
Message-id: <20100706222839.1033.50206.stgit@localhost.localdomain>
Patchwork-id: 10514
O-Subject: [RHEL6.0 qemu-kvm PATCH 11/17] ramblocks: Make use of DeviceState
	pointer and BusInfo.get_dev_path
Bugzilla: 596328
RH-Acked-by: Juan Quintela <quintela@redhat.com>
RH-Acked-by: Amit Shah <amit.shah@redhat.com>
RH-Acked-by: Zachary Amsden <zamsden@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

Bugzilla: 596328
Upstream commit: cc9e98cb8f20d5ef87290591a8e4324c482f3cdd

With these two pieces in place, we can start naming ramblocks.  When
the device is present and it lives on a bus that provides a device
path, we concatenate the path and the provided name.  Otherwise we
just use name.  The resulting id string must be unique.  For now we
assume an allocation for the same name and size is a device that has
been removed and reinserted and return the same block.  This will go
away once qemu_ram_free() is implemented.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>
Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>
---

 cpu-all.h |    1 +
 exec.c    |   29 +++++++++++++++++++++++++++--
 2 files changed, 28 insertions(+), 2 deletions(-)

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>