Revert "virtio: Enable the PUBLISH_USED feature by default for all devices"
RH-Author: Michael S. Tsirkin <mst@redhat.com>
Message-id: <15d570fe46c8cdb0e080beabba98df7bea2c28bd.1279721307.git.mst@redhat.com>
Patchwork-id: 10895
O-Subject: [PATCH RHEL6.0 1/3] Revert "virtio: Enable the PUBLISH_USED feature
	by default for all devices"
Bugzilla: 616501
RH-Acked-by: Juan Quintela <quintela@redhat.com>
RH-Acked-by: Amit Shah <amit.shah@redhat.com>
RH-Acked-by: Alex Williamson <alex.williamson@redhat.com>

This reverts commit 34bb77bdedba99cf3384d16bacdccf8051aff683.

Signed-off-by: Michael S. Tsirkin <mst@redhat.com>

Part of a series reverting PUBLISH_USED.

PUBLISH_USED was reverted upstream.
Discussions on layout and ABI to use are still ongoing,
and there is high risk layout won't match what we put in RHEL,
so we won't be able to support this going forward.

https://bugzilla.redhat.com/show_bug.cgi?id=616501
https://brewweb.devel.redhat.com/taskinfo?taskID=2614379

Tested: on x86_64

---
 hw/virtio-net.h |    2 ++
 hw/virtio.h     |    4 +---
 2 files changed, 3 insertions(+), 3 deletions(-)

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>