Revert "vhost-net: check PUBLISH_USED in backend"
RH-Author: Michael S. Tsirkin <mst@redhat.com>
Message-id: <640fce35830ffbb8677bd77d14d87af841d1a540.1279721307.git.mst@redhat.com>
Patchwork-id: 10896
O-Subject: [PATCH RHEL6.0 2/3] Revert "vhost-net: check PUBLISH_USED in backend"
Bugzilla: 616501
RH-Acked-by: Juan Quintela <quintela@redhat.com>
RH-Acked-by: Amit Shah <amit.shah@redhat.com>
RH-Acked-by: Alex Williamson <alex.williamson@redhat.com>

This reverts commit 2e8898277a9156927e2f282262e392061f39bc7f.

Signed-off-by: Michael S. Tsirkin <mst@redhat.com>

Part of a series reverting PUBLISH_USED.

PUBLISH_USED was reverted upstream.
Discussions on layout and ABI to use are still ongoing,
and there is high risk layout won't match what we put in RHEL,
so we won't be able to support this going forward.

https://bugzilla.redhat.com/show_bug.cgi?id=616501
https://brewweb.devel.redhat.com/taskinfo?taskID=2614379

Tested: on x86_64

---
 hw/vhost_net.c |    3 ---
 1 files changed, 0 insertions(+), 3 deletions(-)

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>