vhost_dev_unassign_memory() don't assert if removing first entry in list.
RH-Author: Jes Sorensen <Jes.Sorensen@redhat.com>
Message-id: <1279898758-7397-2-git-send-email-Jes.Sorensen@redhat.com>
Patchwork-id: 10947
O-Subject: [RHEL6 qemu-kvm PATCH 1/1] vhost_dev_unassign_memory() don't assert
	if removing first entry in list.
Bugzilla: 617085
RH-Acked-by: Marcelo Tosatti <mtosatti@redhat.com>
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Michael S. Tsirkin <mst@redhat.com>
RH-Acked-by: Alex Williamson <alex.williamson@redhat.com>

From: Jes Sorensen <Jes.Sorensen@redhat.com>

If removing an entry from the list which is fully included in the
region and this is the first entry in the list. In this case 'to' can
go to -1, which is perfectly valid. Don't assert() on this case.

Signed-off-by: Jes Sorensen <Jes.Sorensen@redhat.com>
---
 hw/vhost.c |    1 -
 1 files changed, 0 insertions(+), 1 deletions(-)

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>