Fix segfault in mmio subpage handling code.
RH-Author: Gleb Natapov <gleb@redhat.com>
Message-id: <patch-11083-clone-for-rhel6-rhel6>
Patchwork-id: 11090
O-Subject: [RHEL5.5/RHEL6 CVE-2010-2784 PATCH] Fix segfault in mmio subpage
	handling code.
Bugzilla: 619414
CVE: CVE-2010-2784
RH-Acked-by: Petr Matousek <pmatouse@redhat.com>
RH-Acked-by: Juan Quintela <quintela@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

It is possible that subpage mmio is registered over existing memory
page. When this happens "memory" will have real memory address and not
index into io_mem array so next access to the page will generate
segfault. It is uncommon to have some part of a page to be accessed as
memory and some as mmio, but qemu shouldn't crash even when guest does
stupid things. So lets just pretend that the rest of the page is
unassigned if guest configure part of the memory page as mmio.

RHEL5.5. BZ: 619412
RHEL6 BZ: 619421

Upstream: similar patch was sent there.

Signed-off-by: Gleb Natapov <gleb@redhat.com>
--
			Gleb.

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>