Fix migration with spice enabled.
RH-Author: Gerd Hoffmann <kraxel@redhat.com>
Message-id: <1280840535-14453-1-git-send-email-kraxel@redhat.com>
Patchwork-id: 11156
O-Subject: [RHEL-6 kvm PATCH] Fix migration with spice enabled.
Bugzilla: 618168
RH-Acked-by: Luiz Capitulino <lcapitulino@redhat.com>
RH-Acked-by: Juan Quintela <quintela@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

spice hooks into migrate_fd_cleanup() to tell the spice server that
the migration is finished and whenever it was successful or not.

commit 911f1e1368db72694a6590cf8a70c7efb29a899c changes the ordering
of setting FdMigrationState->state and calling migrate_fd_cleanup(),
which trips up spice because it can't figure whenever the migration
worked or not.

This patch restores the old ordering.  It also moves the
FdMigrationState->state update for the close() failure case into the
migrate_fd_cleanup() function, so spice will see those failures too.

bugzilla: #618168
upstream: n/a

Signed-off-by: Gerd Hoffmann <kraxel@redhat.com>
---
 migration.c |    4 ++--
 1 files changed, 2 insertions(+), 2 deletions(-)

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>