disable guest-provided stats on "info ballon" monitor command
RH-Author: Eduardo Habkost <ehabkost@redhat.com>
Message-id: <20100824162302.GF1893@blackpad.lan.raisama.net>
Patchwork-id: 11488
O-Subject: [RHEL6 qemu-kvm PATCH] disable guest-provided stats on "info
	ballon" monitor command
Bugzilla: 623903
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Juan Quintela <quintela@redhat.com>
RH-Acked-by: Luiz Capitulino <lcapitulino@redhat.com>

This feature depends on the guest being alive and working well, but it
breaks "virsh dump" of crashed guests, and can hang virt-manager (see
BZs below).

The proper solution would be adding a timeout on the stats requests, but
disabling the feature reduces our risk at this point in the schedule.

Bugzilla: 623903
Bugzilla: 626544

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>
---
 hw/virtio-balloon.c |    8 +++++++-
 1 files changed, 7 insertions(+), 1 deletions(-)

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>