fix ksmd.init "status"
RH-Author: Andrea Arcangeli <aarcange@redhat.com>
Message-id: <20100831121425.GY6803@random.random>
Patchwork-id: 11631
O-Subject: [RHEL-6 kvm PATCH] fix ksmd.init "status"
Bugzilla: 570467
RH-Acked-by: Markus Armbruster <armbru@redhat.com>
RH-Acked-by: Juan Quintela <quintela@redhat.com>
RH-Acked-by: Dan Kenigsberg <danken@redhat.com>

Patch is untested but it should close the last issue in bug
https://bugzilla.redhat.com/show_bug.cgi?id=570467

The "ksmd status" should be only an aesthetical issue (especially
considering it's not monitoring a real process) but we'll fix it for
correctness and compliance with fedora specs. (i.e. 6.1 only material
I think)

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>