fix '/' and '|' on russian keymap
RH-Author: Eduardo Habkost <ehabkost@redhat.com>
Message-id: <1285957750-19938-1-git-send-email-ehabkost@redhat.com>
Patchwork-id: 12362
O-Subject: [RHEL6 qemu-kvm PATCH] fix '/' and '|' on russian keymap
Bugzilla: 639437
RH-Acked-by: john cooper <john.cooper@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Alex Williamson <alex.williamson@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>

I can't test it myself, but it was reported to be working, by the bug
reporter. The current keymap really looks wrong, as it had duplicate
definitions for "slash" and "bar", and the plain 0x35 scancode is more
likely to work on all guest OSes than the unusual altgr+0x56 scancode.

Bugzilla: 639437
Upstream status: submitted to qemu-devel

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>
---
 pc-bios/keymaps/ru |    2 --
 1 files changed, 0 insertions(+), 2 deletions(-)

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>