BZ #636494 -- cpu check does not correctly enforce CPUID items (V2)
RH-Author: john cooper <john.cooper@redhat.com>
Message-id: <4CB472E3.4090104@redhat.com>
Patchwork-id: 12547
O-Subject: [RHEL6.1 PATCH] BZ #636494 -- cpu check does not correctly enforce
	CPUID items (V2)
Bugzilla: 636494
RH-Acked-by: Alex Williamson <alex.williamson@redhat.com>
RH-Acked-by: Marcelo Tosatti <mtosatti@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

This tightens up "check" and "enforce" treatment
of VMX and SVM flags such that misuse is detected
and reported correctly.  To do so we add another
class of flag, namely those which may exist on
the host and may be conditionally passed to the guest.
For VMX this is never, for SVM this is if (kvm_nested).
We still want to detect/report these features in the case
they simply don't exist on the host (cross-vendor misuse)
first and perform the restriction check if the first check
passes.

Also model_features_t.cpuid wasn't giving us the correct
cpuid function/register diagnostic, so it has been
changed to a string here as the numeric value isn't used.

Upstream status: will be pushed after rhel verification.
Other upstream submitted patch dependencies are still
queued.  And we're thus far the only known users of this
functionality which prods the inverted approach.

Signed-off-by: john cooper <john.cooper@redhat.com>
---

Signed-off-by: Luiz Capitulino <lcapitulino@redhat.com>