Ignore SRAO MCE if another MCE is being processed
RH-Author: Dean Nelson <dnelson@redhat.com>
Message-id: <20101015211735.4348.60230.send-patch@localhost6.localdomain6>
Patchwork-id: 12689
O-Subject: [RHEL6.1 qemu-kvm PATCH v2 2/5] Ignore SRAO MCE if another MCE is
	being processed
Bugzilla: 585910
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Marcelo Tosatti <mtosatti@redhat.com>
RH-Acked-by: Eduardo Habkost <ehabkost@redhat.com>

Resolves RHBZ 585910.

In common cases, guest SRAO MCE will cause corresponding poisoned page
be un-mapped in host and SIGBUS be sent to QEMU-KVM, then QEMU-KVM
will relay the MCE to guest OS.

But it is possible that the poisoned page is accessed in guest after
un-mapped in host and before MCE is relayed to guest OS. So that, the
SRAR SIGBUS is sent to QEMU-KVM before the SRAO SIGBUS, and if
QEMU-KVM relays them to guest OS one by one, guest system may reset,
because the SRAO MCE may be triggered while the SRAR MCE is being
processed. In fact, the SRAO MCE can be ignored in this situation, so
that the guest system is given opportunity to survive.

Backport of upstream commit:
http://git.kernel.org/?p=virt/kvm/qemu-kvm.git;a=commitdiff;h=fb3ba1008cd9af6c7e03ca1d22ec1a904d9719a9

---

 qemu-kvm.c |   28 ++++++++++++++++++++++++++++
 1 files changed, 28 insertions(+), 0 deletions(-)

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>