QMP: Improve debuggability of the BLOCK_IO_ERROR event
RH-Author: Luiz Capitulino <lcapitulino@redhat.com>
Message-id: <20101104115142.7dafa97a@doriath>
Patchwork-id: 13201
O-Subject: [PATCH v2 RHEL6.0.z/RHEL6.1] QMP: Improve debuggability of the
	BLOCK_IO_ERROR event
Bugzilla: 624607
RH-Acked-by: Markus Armbruster <armbru@redhat.com>
RH-Acked-by: Kevin Wolf <kwolf@redhat.com>
RH-Acked-by: Juan Quintela <quintela@redhat.com>

BZ: 624607
Upstream: N/A

We have a vendor extension in the BLOCK_IO_ERROR event which adds
a 'reason' field, which contains a stringfied errno (Please, check
commit a635efd74e to learn why this was needed).

However, it turns out that this information is not enough for
debugging.

Based on a suggestion from Kevin Wolf, this commit adds a new
json-object member to the event, containing the errno value and an
error string from strerror(). Additionally, an error message is
printed to stderr.

All this new information is meant for human debugging only and should
not be used by libvirt (as stated in the event's documentation entry).

Signed-off-by: Luiz Capitulino <lcapitulino@redhat.com>
---

v2:
- Add errno and error string (as a json-object)
- Update event example in QMP/qmp-events.txt

 QMP/qmp-events.txt |   10 +++++++++-
 block.c            |   16 ++++++++++++++++
 2 files changed, 25 insertions(+), 1 deletions(-)

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>