ide: propagate the required alignment
RH-Author: Christoph Hellwig <chellwig@redhat.com>
Message-id: <1289325254-25070-5-git-send-email-chellwig@redhat.com>
Patchwork-id: 13356
O-Subject: [RHEL6.1 qemu PATCH 5/5] ide: propagate the required alignment
Bugzilla: 608548
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Kevin Wolf <kwolf@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

From: Christoph Hellwig <hch@lst.de>

IDE is a bit ugly in this respect.  For one it doesn't really keep track
of a sector size - most of the protocol is in units of 512 bytes, and we
assume 2048 bytes for CDROMs which is correct most of the time.

Second IDE allocates an I/O buffer long before we know if we're dealing
with a CDROM or not, so increase the alignment for the io_buffer
unconditionally.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Kevin Wolf <kwolf@redhat.com>
Upstream commit: 1b2adf28030ee3a570ba3a22401d44da2b18fe01
BZ: 608548

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>