Bug #625333 - qemu treatment of -nodefconfig and -readconfig problematic for debug
RH-Author: john cooper <john.cooper@redhat.com>
Message-id: <4D02FE52.7050405@redhat.com>
Patchwork-id: 14538
O-Subject: [RHEL6.1 PATCH v3] Bug #625333 - qemu treatment of -nodefconfig and
	-readconfig problematic for debug
Bugzilla: 625333
RH-Acked-by: Amit Shah <amit.shah@redhat.com>
RH-Acked-by: Marcelo Tosatti <mtosatti@redhat.com>
RH-Acked-by: Alex Williamson <alex.williamson@redhat.com>

The prior version of this patch had the side effect of allowing a
-readconfig specified config file to be overridden by a wired-in
default config file.  Although this scenario would not occur in
the context of a libvirt launch, altering semantics relative to
upstream qemu is best avoided.

Details:

The problem is qemu currently has only two discrete scans of CLI
arguments.  Trying to live within that structure while supporting
flags which affect other flags is problematic.  Qemu attempts to
implement flag semantics simultaneous with their parsing.  As such
semantic actions follow the unpredictable sequence of args appearing
on the command like.

This patch therefore enforces consistent ordering of the following
semantics as:

- interpret config file parse modifiers
- conditionally read default qemu configs
- read local cpu config
- read any -readconfig configs
- interpret all other options

Testing:

- The config file processing order above should be in effect.
  Specifically an identical data item in a later config file
  should redefine the same introduced by a prior config file.

- Default and local config files may or may not exist on the
  host.  If they do not exist, they are quietly ignored unless
  verbose handling by "-readconfig ?" is specified.  However if
  any do exist and contain config file format errors, qemu will
  error exit.

- Handling of config files specified by "-readconfig <file>"
  is the same as default/local config files with the addition
  qemu will error exit if <file> cannot be opened for read.

- any sequence of "-readconfig" | "-nodefconfig" | "-cpu" has the
  same effect irrespective of relative location in the command line

Upstream Status:

This patch has seen modification in the process of trying to push it
upstream, which resulted in minor syntax changes to the CLI.  We will
need to accommodate this when we rebase qemu.

Signed-off-by: john cooper <john.cooper@redhat.com>
---

Signed-off-by: Luiz Capitulino <lcapitulino@redhat.com>