QMP: Revamp the Python class example
RH-Author: Luiz Capitulino <lcapitulino@redhat.com>
Message-id: <1294168429-1120-12-git-send-email-lcapitulino@redhat.com>
Patchwork-id: 15681
O-Subject: [PATCH 11/16] QMP: Revamp the Python class example
Bugzilla: 647447
RH-Acked-by: Alon Levy <alevy@redhat.com>
RH-Acked-by: Marcelo Tosatti <mtosatti@redhat.com>
RH-Acked-by: Markus Armbruster <armbru@redhat.com>

This commit simplifies and fixes a number of problems in the Python
QEMUMonitorProtocol example class.

It's almost a rewrite and it DOES BREAK the qmp-shell script (which
is going to be fixed in the next commit).

However, I'm not going to split this in different commits because it
could get up to 10 commits, it's really not worth it for a simple
demo class.

Highlights:

 o TCP sockets support
 o QMP events support
 o Add documentation
 o Fix a number of unhandled errors
 o Simplify methods that send commands to the Monitor

Signed-off-by: Luiz Capitulino <lcapitulino@redhat.com>
(cherry picked from commit 1d00a07de980ecf643c7d7699dfc48e94bec15ae)
---
 QMP/qmp.py |  157 ++++++++++++++++++++++++++++++++++++++++-------------------
 1 files changed, 106 insertions(+), 51 deletions(-)

Signed-off-by: Luiz Capitulino <lcapitulino@redhat.com>