Rewrite mouse handlers to use QTAILQ and to have an activation function
RH-Author: Gerd Hoffmann <kraxel@redhat.com>
Message-id: <1294241382-17988-7-git-send-email-kraxel@redhat.com>
Patchwork-id: 15746
O-Subject: [RHEL-6 kvm PATCH 06/38] Rewrite mouse handlers to use QTAILQ and to
	have an activation function
Bugzilla: 642131 634153 615947 632458 631832 647865
RH-Acked-by: Uri Lublin <uril@redhat.com>
RH-Acked-by: Alon Levy <alevy@redhat.com>
RH-Acked-by: Hans de Goede <hdegoede@redhat.com>

And convert usb-hid to use it (to avoid regression with bisection)

Right now, when we do info mice and we've added a usb tablet, we don't see it
until the guest starts using the tablet.  We implement this behavior in order
to provide a means to delay registration of a mouse handler since we treat
the last registered handler as the current handler.

This is a usability problem though as we would like to give the user feedback
that they've either 1) not added an absolute device 2) there is an absolute
device but the guest isn't using it 3) we have an absolute device and it's
active.

By using QTAILQ and having an explicit activation function that moves the
handler to the front of the queue, we can implement the same semantics as
before with respect to automatically switching to usb tablet while providing
the user with a whole lot more information.

upstream: 6fef28ee6e5e0a443857e67aa026d49b6bbdc1b6

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>
Signed-off-by: Gerd Hoffmann <kraxel@redhat.com>
---
 console.h    |    6 +++-
 hw/usb-hid.c |   15 ++++++--
 vl.c         |  109 ++++++++++++++++++++++-----------------------------------
 3 files changed, 58 insertions(+), 72 deletions(-)

Signed-off-by: Luiz Capitulino <lcapitulino@redhat.com>