sdl: use mouse mode notifier
RH-Author: Gerd Hoffmann <kraxel@redhat.com>
Message-id: <1294241382-17988-10-git-send-email-kraxel@redhat.com>
Patchwork-id: 15740
O-Subject: [RHEL-6 kvm PATCH 09/38] sdl: use mouse mode notifier
Bugzilla: 642131 634153 615947 632458 631832 647865
RH-Acked-by: Uri Lublin <uril@redhat.com>
RH-Acked-by: Alon Levy <alevy@redhat.com>
RH-Acked-by: Hans de Goede <hdegoede@redhat.com>

From: Anthony Liguori <aliguori@us.ibm.com>

Today we poll the mouse mode whenever there is a mouse movement.  There is a
subtle usability problem with this though.

If we're in relative mode and grab is enabled, when we change to absolute mode,
we break grab.  This gives a user a seamless transition when the new pointer
is enabled.

But because we poll for mouse change, this grab break won't occur until the user
attempts to move the mouse.  By using notifiers, the grab break happens as soon
as possible.

upstream: 3af12c86d9db42c826dc9a7577b657543baf7fab

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>
Signed-off-by: Gerd Hoffmann <kraxel@redhat.com>
---
 sdl.c |   31 ++++++++++++++++++++-----------
 1 files changed, 20 insertions(+), 11 deletions(-)

Signed-off-by: Luiz Capitulino <lcapitulino@redhat.com>