vgabios update: handle compatibility with older qemu versions
RH-Author: Gerd Hoffmann <kraxel@redhat.com>
Message-id: <1294393606-1424-5-git-send-email-kraxel@redhat.com>
Patchwork-id: 15942
O-Subject: [RHEL-6 kvm PATCH 4/4] vgabios update: handle compatibility with
	older qemu versions
Bugzilla: 638468
RH-Acked-by: Amit Shah <amit.shah@redhat.com>
RH-Acked-by: Alex Williamson <alex.williamson@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

As pointed out by avi the vgabios update is guest-visible and thus has
migration implications.

One change is that the vga has a valid pci rom bar now.  We already have
a pci bus property to enable/disable the rom bar and we'll load the bios
via fw_cfg as fallback for the no-rom-bar case.  So we just have to add
compat properties to handle this case.

A second change is that the magic bochs lfb @ 0xe0000000 is gone.  When
live-migrating a guest from a older qemu version it might be using the
lfb though, so we have to keep it for the old machine types.  The patch
enables the bochs lfb in case we don't have the pci rom bar enabled
(i.e. we are in 0.13+older compat mode).

[ rhel6 backport note: add the compat properties to the new
                       rhel6.0.0 machine type instead ]

upstream: 281a26b15b4adcecb8604216738975abd754bea8

Signed-off-by: Gerd Hoffmann <kraxel@redhat.com>
Cc: avi@redhat.com
Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>
---
 hw/pc.c         |   12 ++++++++++++
 hw/vga-pci.c    |    6 ++++++
 hw/vmware_vga.c |    5 +++++
 3 files changed, 23 insertions(+), 0 deletions(-)

Signed-off-by: Luiz Capitulino <lcapitulino@redhat.com>