device-assignment: Allow PCI to manage the option ROM
RH-Author: Alex Williamson <alex.williamson@redhat.com>
Message-id: <20110117205654.25511.56832.stgit@s20.home>
Patchwork-id: 16424
O-Subject: [RHEL6.1 qemu-kvm PATCH 2/2] device-assignment: Allow PCI to manage
	the option ROM
Bugzilla: 667188
RH-Acked-by: Michael S. Tsirkin <mst@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Don Dutile <ddutile@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=667188
Upstream commit: 576af1b5f5524ca0d6587ae43e395313c14059bb
Brew build: https://brewweb.devel.redhat.com/taskinfo?taskID=3040529

We don't need to duplicate PCI code for mapping and managing the
option ROM for an assigned device.  We're already using an in-memory
copy of the ROM, so we can simply fill the contents from the physical
device and pass the rest off to PCI.  As a benefit, we can now make
use of the rombar and romfile options, which allow us to either hide
the ROM BAR, or load it from an external file, such as we can do
with emulated devices.  This is useful if you want to pass through
and boot from devices that are either missing a physical option ROM
or don't supply a valid option ROM.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>
Signed-off-by: Marcelo Tosatti <mtosatti@redhat.com>
---

 hw/device-assignment.c |  146 ++++++++++++++++++++++--------------------------
 hw/device-assignment.h |    4 +
 2 files changed, 69 insertions(+), 81 deletions(-)

Signed-off-by: Luiz Capitulino <lcapitulino@redhat.com>