remove -redhat-disable-THP
RH-Author: Andrea Arcangeli <aarcange@redhat.com>
Message-id: <20110121055912.GC9506@random.random>
Patchwork-id: 16725
O-Subject: [PATCH RHEL6 qemu-kvm] remove -redhat-disable-THP
Bugzilla: 635418
RH-Acked-by: Luiz Capitulino <lcapitulino@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Eduardo Habkost <ehabkost@redhat.com>
Bugzilla-related: 635418

This is against qemu-kvm origin/rhel60/next as described here:

https://bugzilla.redhat.com/show_bug.cgi?id=635418#c6

-redhat-disable-THP isn't needed because we expect this below patch to
be merged into the RHEL6.1 kernel to enable KSM over THP:

https://bugzilla.redhat.com/show_bug.cgi?id=647334#c3

THP is merged upstream in 2.6.38-rc1 so we could add
madvise(MADV_*HUGEPAGE) to the RHEL6.1 kernel with the right value
(and retain a -redhat-disable-THP), but it's not needed anymore and we
should instead merge the KSM over THP feature as described in the
above in the above link in bug 647334.

I'll be posting soon a patch for qemu upstream but it won't be called
-redhat-disable-THP and it'll be slightly more complex (plus uncertain
how it will change later upstream). There was apparently no problem in
adding a disable-KSM feature in qemu upstream when I asked, just the
command line is uncertain so going with the -redhat-* prefix makes its
safer for now.

===
Subject: remove -redhat-disable-THP

From: Andrea Arcangeli <aarcange@redhat.com>

KSM over THP works so there's no need of this.

Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>
---

Signed-off-by: Luiz Capitulino <lcapitulino@redhat.com>