char: mark socket closed if write fails with -EPIPE
RH-Author: Amit Shah <amit.shah@redhat.com>
Message-id: <d74ff135056165a8d9de68f1a392276e3750505b.1296131377.git.amit.shah@redhat.com>
Patchwork-id: 17135
O-Subject: [RHEL 6.1 PATCH v2 2/2] char: mark socket closed if write fails with
	-EPIPE
Bugzilla: 621484
RH-Acked-by: Alex Williamson <alex.williamson@redhat.com>
RH-Acked-by: Alon Levy <alevy@redhat.com>
RH-Acked-by: Gerd Hoffmann <kraxel@redhat.com>

The current code relies on a read() attempt timing out instead of doing
this.  This is Linux-specific code, so I'm not submitting it upstream.
Upstream will hopefully get a poll() call instead of select(), due to
which this can be handled in main_loop_wait() itself.

Signed-off-by: Amit Shah <amit.shah@redhat.com>
---
 qemu-char.c |    9 ++++++++-
 1 files changed, 8 insertions(+), 1 deletions(-)

Signed-off-by: Luiz Capitulino <lcapitulino@redhat.com>