device-assignment: Cap number of devices we can have assigned
RH-Author: Alex Williamson <alex.williamson@redhat.com>
Message-id: <20110128231209.29293.34356.stgit@s20.home>
Patchwork-id: 17257
O-Subject: [PATCH RHEL6.1 qemu-kvm] device-assignment: Cap number of devices we
	can have assigned
Bugzilla: 670787
RH-Acked-by: Chris Wright <chrisw@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Juan Quintela <quintela@redhat.com>

Upstream status: RHEL specific
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=670787
Brew: https://brewweb.devel.redhat.com/taskinfo?taskID=3073220

This is a RHEL specific patch while we work on fixing this properly
upstream.  We have a fixed number of kvm memory slots, and assigned
device MMIO BARs are a user of those slots.  Slots aren't allocated
until the device is already attached and the guest attempts to
program it, leaving us few options.  The current code will exit when
this occurs.  The only alternative is to use the slow mapping path,
which results in unexpected slownness for a device that's supposed
to be fast.  While we pursue improvements upstream, let's prevent
users from running into this problem by putting a cap on the number
of devices we support.  MAX_DEV_ASSIGN_CMDLINE (8) is the old limit
for the -pcidevice option.  This is hopefully large enough that we
don't hit it casually.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>
---

 hw/device-assignment.c |   15 +++++++++++++++
 1 files changed, 15 insertions(+), 0 deletions(-)

Signed-off-by: Luiz Capitulino <lcapitulino@redhat.com>