virtio-serial: Add rhel6.0.0 compat property for flow control
RH-Author: Amit Shah <amit.shah@redhat.com>
Message-id: <54e7d8fa7ab562afc4907dc17f2e4623540c3892.1296806194.git.amit.shah@redhat.com>
Patchwork-id: 17712
O-Subject: [RHEL6.1 qemu PATCH v5 11/19] virtio-serial: Add rhel6.0.0 compat
	property for flow control
Bugzilla: 588916
RH-Acked-by: Alon Levy <alevy@redhat.com>
RH-Acked-by: Juan Quintela <quintela@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Michael S. Tsirkin <mst@redhat.com>

Add a compat property for the rhel6.0.0 machine type.  When this is used
(via -M rhel6.0.0, for example), the new flow control mechanisms will not
be used.  This is done to keep migration from a machine started with the
rhel600 type on a rhel6.1 installation to a rhel6.0 machine working.

The property is named 'flow_control' and defaults to on.

Signed-off-by: Amit Shah <amit.shah@redhat.com>
---
 hw/pc.c                |    4 +++
 hw/virtio-console.c    |   14 ++++++++++++-
 hw/virtio-pci.c        |    2 +
 hw/virtio-serial-bus.c |   50 ++++++++++++++++++++++++++++++++++++++++++++++++
 hw/virtio-serial.h     |   12 +++++++++++
 5 files changed, 81 insertions(+), 1 deletions(-)

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>